+++ dev-test/githunt/__generated__/repo-info.fragment.stencil-component.tsx
@@ -6,9 +6,9 @@ declare global {
     createdAt: number;
     repository: {
       __typename?: 'Repository';
-      description?: string | null | undefined;
+      description?: string | null;
       stargazers_count: number;
-      open_issues_count?: number | null | undefined;
+      open_issues_count?: number | null;
     };
     postedBy: { __typename?: 'User'; html_url: string; login: string };
   };