+++ dev-test/githunt/types.immutableTypes.ts
@@ -172,16 +172,13 @@ export type OnCommentAddedSubscriptionVariables = Exact<{
 
 export type OnCommentAddedSubscription = {
   readonly __typename?: 'Subscription';
-  readonly commentAdded?:
-    | {
-        readonly __typename?: 'Comment';
-        readonly id: number;
-        readonly createdAt: number;
-        readonly content: string;
-        readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
-      }
-    | null
-    | undefined;
+  readonly commentAdded?: {
+    readonly __typename?: 'Comment';
+    readonly id: number;
+    readonly createdAt: number;
+    readonly content: string;
+    readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
+  } | null;
 };
 
 export type CommentQueryVariables = Exact<{
@@ -192,39 +189,29 @@ export type CommentQueryVariables = Exact<{
 
 export type CommentQuery = {
   readonly __typename?: 'Query';
-  readonly currentUser?:
-    | { readonly __typename?: 'User'; readonly login: string; readonly html_url: string }
-    | null
-    | undefined;
-  readonly entry?:
-    | {
-        readonly __typename?: 'Entry';
-        readonly id: number;
-        readonly createdAt: number;
-        readonly commentCount: number;
-        readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
-        readonly comments: ReadonlyArray<
-          | {
-              readonly __typename?: 'Comment';
-              readonly id: number;
-              readonly createdAt: number;
-              readonly content: string;
-              readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
-            }
-          | null
-          | undefined
-        >;
-        readonly repository: {
-          readonly __typename?: 'Repository';
-          readonly description?: string | null | undefined;
-          readonly open_issues_count?: number | null | undefined;
-          readonly stargazers_count: number;
-          readonly full_name: string;
-          readonly html_url: string;
-        };
-      }
-    | null
-    | undefined;
+  readonly currentUser?: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string } | null;
+  readonly entry?: {
+    readonly __typename?: 'Entry';
+    readonly id: number;
+    readonly createdAt: number;
+    readonly commentCount: number;
+    readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
+    readonly comments: ReadonlyArray<{
+      readonly __typename?: 'Comment';
+      readonly id: number;
+      readonly createdAt: number;
+      readonly content: string;
+      readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
+    } | null>;
+    readonly repository: {
+      readonly __typename?: 'Repository';
+      readonly description?: string | null;
+      readonly open_issues_count?: number | null;
+      readonly stargazers_count: number;
+      readonly full_name: string;
+      readonly html_url: string;
+    };
+  } | null;
 };
 
 export type CommentsPageCommentFragment = {
@@ -239,10 +226,7 @@ export type CurrentUserForProfileQueryVariables = Exact<{ [key: string]: never }
 
 export type CurrentUserForProfileQuery = {
   readonly __typename?: 'Query';
-  readonly currentUser?:
-    | { readonly __typename?: 'User'; readonly login: string; readonly avatar_url: string }
-    | null
-    | undefined;
+  readonly currentUser?: { readonly __typename?: 'User'; readonly login: string; readonly avatar_url: string } | null;
 };
 
 export type FeedEntryFragment = {
@@ -255,10 +239,10 @@ export type FeedEntryFragment = {
     readonly __typename?: 'Repository';
     readonly full_name: string;
     readonly html_url: string;
-    readonly description?: string | null | undefined;
+    readonly description?: string | null;
     readonly stargazers_count: number;
-    readonly open_issues_count?: number | null | undefined;
-    readonly owner?: { readonly __typename?: 'User'; readonly avatar_url: string } | null | undefined;
+    readonly open_issues_count?: number | null;
+    readonly owner?: { readonly __typename?: 'User'; readonly avatar_url: string } | null;
   };
   readonly vote: { readonly __typename?: 'Vote'; readonly vote_value: number };
   readonly postedBy: { readonly __typename?: 'User'; readonly html_url: string; readonly login: string };
@@ -272,32 +256,25 @@ export type FeedQueryVariables = Exact<{
 
 export type FeedQuery = {
   readonly __typename?: 'Query';
-  readonly currentUser?: { readonly __typename?: 'User'; readonly login: string } | null | undefined;
-  readonly feed?:
-    | ReadonlyArray<
-        | {
-            readonly __typename?: 'Entry';
-            readonly id: number;
-            readonly commentCount: number;
-            readonly score: number;
-            readonly createdAt: number;
-            readonly repository: {
-              readonly __typename?: 'Repository';
-              readonly full_name: string;
-              readonly html_url: string;
-              readonly description?: string | null | undefined;
-              readonly stargazers_count: number;
-              readonly open_issues_count?: number | null | undefined;
-              readonly owner?: { readonly __typename?: 'User'; readonly avatar_url: string } | null | undefined;
-            };
-            readonly vote: { readonly __typename?: 'Vote'; readonly vote_value: number };
-            readonly postedBy: { readonly __typename?: 'User'; readonly html_url: string; readonly login: string };
-          }
-        | null
-        | undefined
-      >
-    | null
-    | undefined;
+  readonly currentUser?: { readonly __typename?: 'User'; readonly login: string } | null;
+  readonly feed?: ReadonlyArray<{
+    readonly __typename?: 'Entry';
+    readonly id: number;
+    readonly commentCount: number;
+    readonly score: number;
+    readonly createdAt: number;
+    readonly repository: {
+      readonly __typename?: 'Repository';
+      readonly full_name: string;
+      readonly html_url: string;
+      readonly description?: string | null;
+      readonly stargazers_count: number;
+      readonly open_issues_count?: number | null;
+      readonly owner?: { readonly __typename?: 'User'; readonly avatar_url: string } | null;
+    };
+    readonly vote: { readonly __typename?: 'Vote'; readonly vote_value: number };
+    readonly postedBy: { readonly __typename?: 'User'; readonly html_url: string; readonly login: string };
+  } | null> | null;
 };
 
 export type SubmitRepositoryMutationVariables = Exact<{
@@ -306,7 +283,7 @@ export type SubmitRepositoryMutationVariables = Exact<{
 
 export type SubmitRepositoryMutation = {
   readonly __typename?: 'Mutation';
-  readonly submitRepository?: { readonly __typename?: 'Entry'; readonly createdAt: number } | null | undefined;
+  readonly submitRepository?: { readonly __typename?: 'Entry'; readonly createdAt: number } | null;
 };
 
 export type RepoInfoFragment = {
@@ -314,9 +291,9 @@ export type RepoInfoFragment = {
   readonly createdAt: number;
   readonly repository: {
     readonly __typename?: 'Repository';
-    readonly description?: string | null | undefined;
+    readonly description?: string | null;
     readonly stargazers_count: number;
-    readonly open_issues_count?: number | null | undefined;
+    readonly open_issues_count?: number | null;
   };
   readonly postedBy: { readonly __typename?: 'User'; readonly html_url: string; readonly login: string };
 };
@@ -328,16 +305,13 @@ export type SubmitCommentMutationVariables = Exact<{
 
 export type SubmitCommentMutation = {
   readonly __typename?: 'Mutation';
-  readonly submitComment?:
-    | {
-        readonly __typename?: 'Comment';
-        readonly id: number;
-        readonly createdAt: number;
-        readonly content: string;
-        readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
-      }
-    | null
-    | undefined;
+  readonly submitComment?: {
+    readonly __typename?: 'Comment';
+    readonly id: number;
+    readonly createdAt: number;
+    readonly content: string;
+    readonly postedBy: { readonly __typename?: 'User'; readonly login: string; readonly html_url: string };
+  } | null;
 };
 
 export type VoteButtonsFragment = {
@@ -353,13 +327,10 @@ export type VoteMutationVariables = Exact<{
 
 export type VoteMutation = {
   readonly __typename?: 'Mutation';
-  readonly vote?:
-    | {
-        readonly __typename?: 'Entry';
-        readonly score: number;
-        readonly id: number;
-        readonly vote: { readonly __typename?: 'Vote'; readonly vote_value: number };
-      }
-    | null
-    | undefined;
+  readonly vote?: {
+    readonly __typename?: 'Entry';
+    readonly score: number;
+    readonly id: number;
+    readonly vote: { readonly __typename?: 'Vote'; readonly vote_value: number };
+  } | null;
 };