+++ dev-test/gql-tag-operations-urql/gql/graphql.ts
@@ -107,21 +107,15 @@ export type User = {
 
 export type FooQueryVariables = Exact<{ [key: string]: never }>;
 
-export type FooQuery = {
-  __typename?: 'Query';
-  Tweets?: Array<{ __typename?: 'Tweet'; id: string } | null | undefined> | null | undefined;
-};
+export type FooQuery = { __typename?: 'Query'; Tweets?: Array<{ __typename?: 'Tweet'; id: string } | null> | null };
 
-export type LelFragment = { __typename?: 'Tweet'; id: string; body?: string | null | undefined };
+export type LelFragment = { __typename?: 'Tweet'; id: string; body?: string | null };
 
 export type BarQueryVariables = Exact<{ [key: string]: never }>;
 
 export type BarQuery = {
   __typename?: 'Query';
-  Tweets?:
-    | Array<{ __typename?: 'Tweet'; id: string; body?: string | null | undefined } | null | undefined>
-    | null
-    | undefined;
+  Tweets?: Array<{ __typename?: 'Tweet'; id: string; body?: string | null } | null> | null;
 };
 
 export const LelFragmentDoc = {