+++ dev-test/star-wars/__generated__/HeroDetails.tsx
@@ -2,7 +2,7 @@ import * as Types from '../types.d';
 
 import { gql } from '@apollo/client';
 import * as Apollo from '@apollo/client';
-const defaultOptions = {};
+const defaultOptions = {} as const;
 export type HeroDetailsQueryVariables = Types.Exact<{
   episode?: Types.InputMaybe<Types.Episode>;
 }>;
@@ -10,10 +10,9 @@ export type HeroDetailsQueryVariables = Types.Exact<{
 export type HeroDetailsQuery = {
   __typename?: 'Query';
   hero?:
-    | { __typename?: 'Droid'; primaryFunction?: string | null | undefined; name: string }
-    | { __typename?: 'Human'; height?: number | null | undefined; name: string }
-    | null
-    | undefined;
+    | { __typename?: 'Droid'; primaryFunction?: string | null; name: string }
+    | { __typename?: 'Human'; height?: number | null; name: string }
+    | null;
 };
 
 export const HeroDetailsDocument = gql`