+++ dev-test/test-schema/types.preResolveTypes.onlyOperationTypes.ts
@@ -16,7 +16,7 @@ export type TestQueryVariables = Exact<{ [key: string]: never }>;
 
 export type TestQuery = {
   __typename?: 'Query';
-  testArr1?: Array<string | null | undefined> | null | undefined;
-  testArr2: Array<string | null | undefined>;
+  testArr1?: Array<string | null> | null;
+  testArr2: Array<string | null>;
   testArr3: Array<string>;
 };