+++ packages/plugins/typescript/urql/tests/urql.spec.ts
@@ -552,13 +552,61 @@ query MyFeed {
       )) as Types.ComplexPluginOutput;
 
       expect(content.content).toBeSimilarStringTo(`
-export function useFeedQuery(options: Omit<Urql.UseQueryArgs<FeedQueryVariables>, 'query'> = {}) {
+export function useFeedQuery(options?: Omit<Urql.UseQueryArgs<FeedQueryVariables>, 'query'>) {
   return Urql.useQuery<FeedQuery>({ query: FeedDocument, ...options });
 };`);
 
       expect(content.content).toBeSimilarStringTo(`
 export function useSubmitRepositoryMutation() {
   return Urql.useMutation<SubmitRepositoryMutation, SubmitRepositoryMutationVariables>(SubmitRepositoryDocument);
+};`);
+      await validateTypeScript(content, schema, docs, {});
+    });
+
+    it('Should generate hooks for query with required arguments', async () => {
+      const documents = parse(/* GraphQL */ `
+        query RequiredArg($feedType: FeedType!, $limit: Int = 10) {
+          feed(type: $feedType, limit: $limit) {
+            id
+          }
+        }
+      `);
+      const docs = [{ location: '', document: documents }];
+
+      const content = (await plugin(
+        schema,
+        docs,
+        { withHooks: true, withComponent: false },
+        { outputFile: 'graphql.tsx' }
+      )) as Types.ComplexPluginOutput;
+
+      expect(content.content).toBeSimilarStringTo(`
+export function useRequiredArgQuery(options: Omit<Urql.UseQueryArgs<RequiredArgQueryVariables>, 'query'>) {
+  return Urql.useQuery<RequiredArgQuery>({ query: RequiredArgDocument, ...options });
+};`);
+      await validateTypeScript(content, schema, docs, {});
+    });
+
+    it('Should generate hooks for query with default value arguments', async () => {
+      const documents = parse(/* GraphQL */ `
+        query DefaultValueArg($feedType: FeedType! = "HOT", $limit: Int) {
+          feed(type: $feedType, limit: $limit) {
+            id
+          }
+        }
+      `);
+      const docs = [{ location: '', document: documents }];
+
+      const content = (await plugin(
+        schema,
+        docs,
+        { withHooks: true, withComponent: false },
+        { outputFile: 'graphql.tsx' }
+      )) as Types.ComplexPluginOutput;
+
+      expect(content.content).toBeSimilarStringTo(`
+export function useDefaultValueArgQuery(options?: Omit<Urql.UseQueryArgs<DefaultValueArgQueryVariables>, 'query'>) {
+  return Urql.useQuery<DefaultValueArgQuery>({ query: DefaultValueArgDocument, ...options });
 };`);
       await validateTypeScript(content, schema, docs, {});
     });
@@ -635,6 +683,31 @@ export function useSubmitRepositoryMutation() {
       )) as Types.ComplexPluginOutput;
 
       expect(content.content).toContain(`export function useTest(`);
+    });
+
+    it('should respect dedupeOperationSuffix for hooks', async () => {
+      const docs = [
+        {
+          location: '',
+          document: parse(/* GraphQL */ `
+            query testQuery {
+              feed {
+                id
+              }
+            }
+          `),
+        },
+      ];
+      const content = (await plugin(
+        schema,
+        docs,
+        { withHooks: true, dedupeOperationSuffix: true },
+        {
+          outputFile: 'graphql.tsx',
+        }
+      )) as Types.ComplexPluginOutput;
+
+      expect(content.content).toContain(`export function useTestQuery(`);
     });
 
     it('should output warning if documentMode = external and importDocumentNodeExternallyFrom is not set', async () => {