+++ packages/graphql-codegen-cli/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @graphql-codegen/cli
 
+## 2.4.0
+
+### Minor Changes
+
+- 4c42e2a71: Performance optimizations in schema and documents loading (shared promises)
+
+## 2.3.1
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [8643b3bf3]
+- Updated dependencies [b61dc57cf]
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/core@2.4.0
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 2.3.0
 
 ### Minor Changes