+++ packages/plugins/other/visitor-plugin-common/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @graphql-codegen/visitor-plugin-common
 
+## 2.5.2
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 2.5.1
 
 ### Patch Changes