+++ packages/plugins/typescript/operations/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @graphql-codegen/typescript-operations
 
+## 2.2.3
+
+### Patch Changes
+
+- 7649201fe: Remove redundant mandatory 'undefined' on avoidOptionals=true
+
+## 2.2.2
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/visitor-plugin-common@2.5.2
+  - @graphql-codegen/typescript@2.4.2
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 2.2.1
 
 ### Patch Changes