+++ packages/plugins/typescript/urql/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @graphql-codegen/typescript-urql
 
+## 3.5.0
+
+### Minor Changes
+
+- 508e078aa: feat: Improve query hooks arg types
+
+### Patch Changes
+
+- ca6132e25: Respect dedupeOperationSuffix in hooks
+
+## 3.4.2
+
+### Patch Changes
+
+- 6002feb3d: Fix exports in package.json files for react-native projects
+- Updated dependencies [6002feb3d]
+  - @graphql-codegen/visitor-plugin-common@2.5.2
+  - @graphql-codegen/plugin-helpers@2.3.2
+
 ## 3.4.1
 
 ### Patch Changes